requirements.txt
Pillow==8.3.1
black==21.8b0
pylint==2.10.2
isort==5.9.3
pre-commit==2.15.0
